Add loading efficiency calculation
A difference remains between this calculation and the R calculation. The R code
is included as a comment, as well as some commented out debug code. The actual
analysis code post "z" appears to be correct, and the contents of "z" appear to
be correct, but the two together are off by a small amount.

A failing unit test with some hardcoded paths is included to aid development.